Rename input methods
It is not a concern of the game logic to handler user input. It is the
UI’s responsibility to translate the input to the commands for tha game.

The user inputs a direction. The UI determines it is a command for the
snake to slither. It should call the appropriate game’s slither command.

Renamed input_* to slither_* to better represent the concern split.